Add necessary Icons for Deductions Tab. Implement 'Delete Deduction' function. Breakdown components of Collapsible Table components. Destructured imports.
